Interpret element always as reference type
To solve the issue with abstract elements we need a way to distinguish
between an element and it's actual type. In the old implementation types
inlined inside elements will be interpreted directly as type of the
element. With the new implementation we add an extra layer on
indirection by forcing the  element to be always a reference. That gives
us the ability to change the element from a reference type to an
abstract type if the element is used as substitution group head.