Refactor resource resolvers
The old approach was to represent any resource that needs to be resolved
by an suitable `Url` object, but URLs do not support relative paths, so
we switched the behavior to keep the location of the resource to resolve
as string and then then resolver needs to find a suitable interpretation
for it. In most cases we can simply use `Url::join` to make a valid URL
from the URL of the current processed schema and the location of the one.