Remove the content mode parameter from the generator
The content mode was used to tell the generator if a struct or enum
should be generated for a complex type. This option lead to an unneeded
complexity in the generator. So we removed this option. Now the generator
will always generate enums for choices and structs for sequences. If the
conversion between structs and enums is still needed, a suitable method
could be implemented in the optimizer.