Fix content handling in the code generator
The handling of the content of a XML element was not correct, resulting
in incorrect code. The content can have a cardinality too, so that it
might need a container type (like Vec or Option). The content can
only be flattened if it has a cardinality of one.

Caution: This commit fixes only the code that is generated for the type.
The code generated for quick_xml de-/serialization is not updated,
which result in broken integration tests and will be fixed in a separate
commit.