Refactor the generator types
The old structure of the "data" types in the generator mixed information
of the actual type that should be generated and the original type load
from the schema. This commit refactors the whole types implemented in
the generator, that they now reflect just the type that should be
generated. This makes it much more easier to write the actual render
code, because we don't have to deal with additional checks inside the
renderer. It basically converts the type form the schema to a type that
contains only render related information.

Caution: `quick_xml` related code generation and unit tests are still
broken.