Fix XML schema parsing
The XML schema contains HTML inside a `xs:documentation`. In the old
version we returned the events related to the HTML code to the parent
deserializer for additional processing, which is wrong if the current
deserializer is not finished yet. So we added an additional type called
`DeserializerEvent` to indicate what to do with an event returned by a
deserializer. This is needed and can't be done by just raising an error
or dropping the related event, because the root of the deserialization
process is used to skip whole XML sub-trees if needed, so the event
needs to be returned to the root.