Improve name generation
It was unclear which of the `Name` values in the code were generated and
which were just applied from the schema. So the `Name` now has only two
states: `Named` and `Generated`. The actual generation of the name is
moved to a new `NameBuilder` type.

The new name type makes it possible to generate code for all types, for
all named types or for a set of user defined names. This makes it easier
to control for which types code is actually generated.