Add support for `xs:anySimpleType` and fix interpreter bug
The interpreter had a bug, that caused complex types with simple content
to be interpreted as simple types, which resulted in a panic, because the
interpretation of the simple type then tried to change the fixed variant
of the builder. To fix that we set the type mode to simple only if it is
not known yet, this will then clone the content type for manipulation in
a later code path.