Fix deserialization of `bool` values
According to the XML specification boolean values can be `true` or `false`,
but also `1` or `0`. This caused an error in the deserialization because we
used the `FromStr` implementation of `bool` to deserialize the value.

Any type that implemented `FromStr` automatically got an blanked implementation
for `DeserializeBytes`. To solve the issue with custom values, we moved the blanked
implementation to a new `DeserializeBytesFromStr` marker trait, a types needs to
be marked with to use the `FromStr` implementation and then implemented `DeserializeBytes`
for `bool` manually

The changes were also applied analog to `SerializeBytes`, `ToString` and `SerializeBytesToString`.