Refactor render step configuration
This commit adds a new `RenderStepConfig` trait that helps to deal with
the tradeoffs the config should respect corresponding to the different
render steps. For example it handles cloning the render step config and
deals with the detection of mutual exclusive steps.
The new trait has a default implementation for all `RenderSteps` that
also implement `Clone` and for the existing `config::RenderStep` type.
This enables the user to add either `config::RenderSteps` or any type
that implements the `RenderStep` trait to the config, making the public
interface of the config more flexible.