Fix issue with quick_xml sequence deserialization
During deserialization of a sequence, the deserializer moved on to the
next element, if the current event did not match the name of the current
expected element, event if the number of expected elements was not fulfilled.

To fix this we now call then corresponding handle method for that element
in any case using a suitable `DeserializerOutput` that indicates to continue
parsing the current element.

As a cool side effect, this also reduces the general size of the generated
code :)