Implement configuration for the `quick_xml` serializer
During serialization different options exists for dealing with the
namespaces of an element. The current implementation just added each
namespace using a suitable namespace prefix and then also prefixed each
emitted element. This may reduce the overhead of the resulting XML.
To fix this we introduced a config for the `quick_xml` serializer to
configure how exactly the namespaces are handled during serialization.