Fix error included schemas without target namespace
When other schemas are included into another schema, while both of them
do not specify a target namespace, the elements defined in the included
schema could not be resolved inside the interpreter. To solve this we
now additionally try to lookup the element in all other schemas of the
current "unknown" namespace by using the the namespace id of the current
processed schema.