Improve loop detection
The old approach of the loop detection was to check if a type could cause
a loop by traversing the whole type tree with the type to render as root.
This took a lot of time for large schemas with lot of type dependencies.

The new approach is, to keep track of the types, while they are rendered
using simple type ids and bit sets. This increases the performance of the
generator by factor 10! The only downside compared to the old approach is,
that now the leaves in the type tree are boxed if needed instead of the
root.