Fix naming error in `quick_xml` de-/serializer
In one of the later commits (see `7a3cf007`) we've implemented a fix for
naming conflicts between element references with the same name from
different namespaces. The fix included to change the identifier of the
conflicting elements, which caused problems during the de/-serialization
of that element. The correct approach would be to set a default `display_name`
for the conflicting elements. This was fixed in this commit.