Use the name that is passed to the serializer for `AnyElement`
`AnyElement` is used in two contexts:
- If the schema defines a `xs:any` element, or
- If the schema defines a explicit element without a concrete type.

While it is correct for the sooner, to not use the name from the element
definition (because it is a generated name like "anyX"), the later should
make use of the correct element name.
Instead of silently dropping the name in the `AnyElement` serializer, we
now use it to serialize the correct element. To avoid conflicts with the
`xs:any` elements the code generator for the serializer does not pass
the name to the serializer of the element was a `xs:any` element.