Unify configuration for XML types
Before this change the configuration for types that handle XML data in a
general or dynamic way, had different approaches. Some of the types were
hardcoded and enabled by a flag (`Text`, `Mixed` and `Nillable`), others
were enabled by passing a suitable type to the config (`AnyElement` and
`AnyAttributes`). This commit unifies the config by providing a enable
flag for all types and a type path with a suitable default value.