Implement support for duplicate types in different schema files
While it is obviously not legal to have different element types with the exact same identifiers in an XML schema, it is possible to tolerate this in xsd-parser. The basic idea is to add a SchemaId to every Ident, therefore allowing otherwise identical Idents to coexist if they were identified in different schema files. This sounds simple, but creates the problem that if schema 1 defines foo:bar and then schema 2 references it, the Ident from the reference would be tagged as schema 2 and would therefore not resolve to the ident from schema 1. So most of the work then boils down to adding fallback logic everywhere so that if an Ident is not found with the same SchemaId it correctly resolves to an Ident defined elsewhere, just like before.