Fix windows build
When referencing the repository directly and not the crate on crates.io
the build script is broken on windows because the symlink to the `doc`
directory inside `xsd-parser` is broken. To fix this we try to resolve
the file in the current crate directory and in it's parent directory
inside the build script.